--- src/breakdown.py
@@ -1,2 +1,0 @@
-dam_line=102,9/18/2024 16:21:09.7962,SPELL_DAMAGE,Player-580-0AB7EBF0,Nevx-Blackmoore-EU,1297,128,Player-2073-09EC6479,Flåppers-Stormscale-EU,66888,0,204197.0,Purge the Wicked,6.0,,0.0,0.0,0.0,0.0,0.0,,20357.0,26923.0,,,0.0,6389283.0,6424940.0,78199.0,0.0,3.0,118.0,140.0,,,
-
