+++ src/run_parser.py
@@ -7,7 +7,7 @@ current_dir = os.path.dirname(os.path.abspath(__file__))
 parent_dir = os.path.abspath(os.path.join(current_dir, ".."))
 sys.path.append(parent_dir)
 
-from wowclp import Parser
+from parser import Parser
 
 if __name__ == "__main__":
     psr = Parser()
